ДЗ № 14: модификация классов и пакетов
